refactor(tests/task): create a test class for each feature
